advanced-data/kafka/requirements.txt
python-dotenv
confluent_kafka